Remove redundant type casting

+++ Torpedo/src/etri/sdn/controller/module/firewall/OFFirewallRuleReplySerializerModule.java
@@ -22,7 +22,7 @@ final class OFFirewallRuleSerializer extends JsonSerializer<FirewallRule> {
 	throws IOException, JsonProcessingException {
 		
 		jgen.writeStartObject();
-		jgen.writeNumberField("ruleid", (Integer)rule.ruleid);
+		jgen.writeNumberField("ruleid", rule.ruleid);
 		jgen.writeNumberField("dpid", rule.dpid);
 		jgen.writeNumberField("in_port", rule.in_port);
 		jgen.writeNumberField("dl_src", rule.dl_src);
